LayoutTests/fast/forms/input-valueasdate-time-expected.txt

Tests for .valueAsDate with <input type=time>.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS valueAsDateFor("") is null
PASS valueAsDateFor("00:00:00.000").getTime() is Date.UTC(1970, 0, 1, 0, 0, 0, 0)
PASS valueAsDateFor("04:35").getTime() is Date.UTC(1970, 0, 1, 4, 35, 0, 0)
PASS valueAsDateFor("23:59:59.999").getTime() is Date.UTC(1970, 0, 1, 23, 59, 59, 999)
PASS successfullyParsed is true

TEST COMPLETE
